improved for loop type checking and lowering

+++ tests/FLang.Tests/Harness/control_flow/for_loop.f
@@ -2,7 +2,7 @@
 //! EXIT: 42
 pub fn main() i32 {
     for (i in 0..3) {
-        let x: i32 = i
+        let x: isize = i
     }
     return 42
 }
